KVM: arm64: Reduce the size of the vcpu flag members

+++ arch/arm64/include/asm/kvm_host.h
@@ -336,13 +336,13 @@ struct kvm_vcpu_arch {
 	} fp_state;
 
 	/* Configuration flags */
-	u64 cflags;
+	u8 cflags;
 
 	/* Input flags to the hypervisor code */
-	u64 iflags;
+	u8 iflags;
 
 	/* State flags, unused by the hypervisor code */
-	u64 sflags;
+	u8 sflags;
 
 	/*
 	 * We maintain more than a single set of debug registers to support
